Add 'drag and drop' ordering for gallery images
jQuery-ui-sortable used to implement drag and drop ordering of image
attachments in the gallery. Order of images is stored in the database
(wp_postmeta). Order of images is then honoured when loading gallery
from both front end (view classified) and back end (edit classified).
Ajax is used to update order as it is changed.